several bug fixes using the new test cases.
Former-commit-id: 36567220a6b658f9dbff8091561e53ed45c4bf7a

+++ src/debug.h
@@ -20,4 +20,4 @@
 
 
 
-void printParticle(struct part *parts, int i);
+void printParticle(struct part *parts, long long int i);
